+++ pathCounts.c
@@ -19,10 +19,10 @@ pathCountsMatrix *countPaths(pathCountsWithPredMatrix *pathCountsWithPred, compa
     pathCounts->nbCols = compact->nbNodes;
     pathCounts->data = mallocOrDie(sizeof(float) * nbNodes * nbNodes, "E: OOM for path counts data\n");
     
-    for (unsigned int i = 0; i < nbNodes; i++) {
-        for (unsigned int j = 0; j < nbNodes; j++) {
+    for (size_t i = 0; i < nbNodes; i++) {
+        for (size_t j = 0; j < nbNodes; j++) {
             float sum = 0;
-            for (unsigned int k = compact->offsets[j]; k < compact->offsets[j + 1]; k++) {
+            for (size_t k = compact->offsets[j]; k < compact->offsets[j + 1]; k++) {
                 sum += pathCountsWithPred->data[i * sumDegrees + k];
             }
             pathCounts->data[i * nbNodes + j] = sum;
@@ -33,8 +33,8 @@ pathCountsMatrix *countPaths(pathCountsWithPredMatrix *pathCountsWithPred, compa
 }
 
 void printPathCounts(pathCountsMatrix *pathCounts) {
-    for (unsigned int i = 0; i < pathCounts->nbCols; i++) {
-        for (unsigned int j = 0; j < pathCounts->nbCols; j++) {
+    for (size_t i = 0; i < pathCounts->nbCols; i++) {
+        for (size_t j = 0; j < pathCounts->nbCols; j++) {
             printf("%0.2f ", pathCounts->data[i * pathCounts->nbCols + j]);
         }
         printf("\n");
